pyproject.toml
[project]
name = "translip"
version = "0.1.0"
description = "Separate voice and background audio from video or audio inputs."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "demucs>=4.0.1,<5",
  "faster-whisper>=1.1.1,<2",
  "imageio-ffmpeg>=0.6.0,<1",
  "librosa>=0.10.2,<1",
  "numpy>=1.26,<3",
  "qwen-tts>=0.1.1",
  "scikit-learn>=1.5,<2",
  "sentencepiece>=0.2,<1",
  "soundfile>=0.12.1,<1",
  "speechbrain>=1.0.3,<2",
  "torchcodec>=0.8.0,<1",
  "transformers>=4.51,<5",
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.32,<1",
  "sqlmodel>=0.0.22,<1",
  "python-multipart>=0.0.12,<1",
  "aiofiles>=24.1,<25",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.24,<1",
  "httpx>=0.27,<1",
]

[project.scripts]
translip = "translip.cli:main"
translip-server = "translip.server.app:run_server"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
